requirements.txt
streamlit
openai
faiss-cpu
pypdf
tqdm
python-dotenv
